int language accuracy

+++ cld2wrapper.h
@@ -31,7 +31,7 @@ extern const short kAvgDeltaOctaScore[];
 typedef struct DetectedLanguage {
     const char* language_name;
     const char* language_code;
-    float language_accuracy;
+    int language_accuracy;
     bool is_reliable;
 } DetectedLanguage;
